+++ remoter/src/main/java/remoter/compiler/builder/ClassBuilder.java
@@ -55,6 +55,7 @@ class ClassBuilder extends RemoteBuilder {
                 .addParameter(ClassName.get("android.os", "IBinder"), "binder")
                 .addStatement("this.mRemote = binder")
                 .addStatement("this._binderID = __getStubID()")
+                .addStatement("this._stubProcess = __getStubProcessID()")
                 .build());